Database/pyproject.toml
[project]
name = "database"
version = "0.1.0"
description = "Database wrapper scripts for the repository"
readme = "README.md"

requires-python = ">=3.12"
dependencies = [
    "firebase-admin>=7.3.0",
    "python-dotenv>=1.2.2",
]

[build-system]
requires = ["uv_build>=0.10.9,<0.11.0"]
build-backend = "uv_build"
